% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulocityr.R
\docType{package}
\name{cumulocityr-package}
\alias{cumulocityr-package}
\alias{_PACKAGE}
\title{R client for the Cumulocity API}
\description{
This package contains functions to interact with the Cumulocity API.
}
\section{Functions}{

\itemize{
\item \code{\link{get_devices}} returns available devices for a tenant.
\item \code{\link{get_measurements}} returns measurements for a device.
}
}

\section{Authentication}{


Environment variables used for authentication are stored in a local .Renviron file:
\itemize{
\item \env{CUMULOCITY_base_url = "[tenant url]"}\cr
\item \env{CUMULOCITY_usr = "[username]"}\cr
\item \env{CUMULOCITY_pwd = "[password]"}\cr
\item \env{CUMULOCITY_device_id = "[an example device id]"}
}

The tenant url should be of the form \code{"https://tenant_name.cumulocity.com"}.
}

\section{Other considerations}{


There is a limit of 2000 on \code{pageSize} for each call.
}

\section{References}{

\itemize{
\item \href{https://cumulocity.com/guides/reference/inventory/}{Cumulocity Inventory API}
\item \href{https://cumulocity.com/guides/reference/events/}{Cumulocity Events API}
\item \href{https://cumulocity.com/guides/reference/measurements/}{Cumulocity Measurements API}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://softwareag.github.io/cumulocityr/}
  \item \url{https://github.com/SoftwareAG/cumulocityr}
  \item Report bugs at \url{https://github.com/SoftwareAG/cumulocityr/issues}
}

}
\author{
\strong{Maintainer}: Dmitriy Bolotov \email{rpmmlsupport@softwareag.com}

Other contributors:
\itemize{
  \item Software AG [copyright holder]
}

}
\keyword{internal}
