% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get the measurements for a device for a time period.}
\usage{
get_data(device_id, date_from = NULL, date_to = NULL, page_size = 10,
  abridged = TRUE, parse_time = TRUE)
}
\arguments{
\item{device_id}{The device id.}

\item{date_from}{The starting datetime.}

\item{date_to}{The ending datetime.}

\item{page_size}{The page size.}

\item{abridged}{If TRUE, exclude "self" and "source" fields from results.}

\item{parse_time}{If TRUE, parse "time" field from char to POSIXlt.}
}
\value{
R object with measurements.
}
\description{
Get the measurements for a device for a time period.
}
\details{
Get the measurements for a device for a time period.

The parameter \code{page_size} is only used if \code{date_from}
or \code{date_to} are NULL, or both are NULL.
}
\examples{
\dontrun{
get_data(device_id)
}
}
\references{
\href{https://cumulocity.com/guides/reference/events/}{Cumulocity Events API}
\href{https://cumulocity.com/guides/reference/measurements/}{Cumulocity Measurements API}
}
\author{
Dmitriy Bolotov
}
