% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_devices.R
\name{get_devices}
\alias{get_devices}
\title{Get the devices or group of devices for a tenant.}
\usage{
get_devices(page_size = 2000, parse_datetime = TRUE,
  parse_json = TRUE)
}
\arguments{
\item{page_size}{The page size, set to 2000 (maximum) by default.}

\item{parse_datetime}{If TRUE, parse datetime fields from char to POSIXlt.}

\item{parse_json}{If TRUE, parse the JSON object into a data frame.}
}
\value{
A \code{data.frame} with measurements.
}
\description{
Get the devices or group of devices for a tenant.
}
\details{
Get the devices or group of devices for a tenant.

If \code{parse_datetime = TRUE}, the following fields are parsed from char to
POSIXlt: \code{creationTime, lastUpdated, c8y_Availability.lastMessage}.

If \code{parse_json} is TRUE, the JSON object is parsed using \code{jsonlite::fromJSON}
before being returned. The data is converted to a single flattened data frame.

If \code{parse_json} is FALSE, the JSON object is returned as a JSON string.
The parameter \code{parse_datetime} has no effect.
}
\examples{
\dontrun{
get_devices()
}

}
\references{
\href{https://cumulocity.com/guides/reference/inventory/}{Cumulocity Inventory API}
}
\author{
Dmitriy Bolotov
}
