% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_events.R
\name{get_events}
\alias{get_events}
\title{Get the events for a device.}
\usage{
get_events(device_id, date_from = NULL, date_to = NULL,
  page_size = 2000, pages_per_query = 1, start_page = 1,
  parse_datetime = TRUE, parse_json = TRUE)
}
\arguments{
\item{device_id}{The device id.}

\item{date_from}{The starting datetime.}

\item{date_to}{The ending datetime.}

\item{page_size}{The page size, set to 2000 (maximum) by default.}

\item{pages_per_query}{The number of pages to return per function call.}

\item{start_page}{The first page used in the query.}

\item{parse_datetime}{If TRUE, parse "time" field from char to POSIXlt.}

\item{parse_json}{If TRUE, parse the JSON object into a data frame.}
}
\value{
A \code{data.frame} with events.


If \code{parse_json} is TRUE, the JSON object is parsed using \code{jsonlite::fromJSON}
before being returned. The data is converted to a single flattened data frame.
If a page does not contain any events, it does not get added to the data frame.

If \code{parse_json} is FALSE, the JSON object is returned as a JSON string. For queries with multiple pages, a
list of such objects is returned. All pages are added to the list, even if there are no events.
The parameter \code{parse_datetime} has no effect.
}
\description{
Get the events for a device.
}
\details{
Get the events for a device for a time period.

The parameter \code{page_size} is only used if \code{date_from}
or \code{date_to} are NULL, or both are NULL.
}
\examples{
\dontrun{
get_events(device_id)
}
}
\references{
\href{https://cumulocity.com/guides/reference/events/}{Cumulocity Events API}
}
\author{
Dmitriy Bolotov
}
