% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_measurements.R
\name{get_measurements}
\alias{get_measurements}
\title{Get the measurements for a device for a time period.}
\usage{
get_measurements(device_id, date_from = NULL, date_to = NULL,
  page_size = 2000, pages_per_query = 1, start_page = 1,
  drop_fields = TRUE, parse_datetime = TRUE, parse = TRUE)
}
\arguments{
\item{device_id}{The device id.}

\item{date_from}{The starting datetime.}

\item{date_to}{The ending datetime.}

\item{page_size}{The page size, set to 2000 (maximum) by default.}

\item{pages_per_query}{The number of pages to return per function call.}

\item{start_page}{The first page used in the query.}

\item{drop_fields}{If TRUE, exclude "self" and "source" fields from the result.}

\item{parse_datetime}{If TRUE, parse "time" field from char to POSIXlt.}

\item{parse}{If TRUE, parse the JSON object into a data frame.}
}
\value{
A \code{data.frame} with measurements.

Note that some columns in the returned data frame contain data frames themselves.
For example, \code{c8y_Mobile} is a \code{data.frame} of 11 variables.

If both \code{date_from} and \code{date_to} are present, \code{page_size}
is not used.

If \code{page_size} and both dates are NULL, the function will return up to 5 rows of data.

If \code{parse} is TRUE, the JSON object is parsed using \code{jsonlite::fromJSON}
before being returned. The data is converted to a single flattened data frame.
If a page does not contain any measurements, it does not get added to the data frame.

If \code{parse} is FALSE, the JSON object is returned as a JSON string. For queries with multiple pages, a
list of such objects is returned. All pages are added to the list, even if there are no measurements.
The params \code{drop_fields} and \code{parse_datetime} have no effect.
}
\description{
Get the measurements for a device for a time period.
}
\details{
Get the measurements for a device for a time period.

The parameter \code{page_size} is only used if \code{date_from}
or \code{date_to} are NULL, or both are NULL.
}
\examples{
\dontrun{
get_measurements(device_id)
}
}
\references{
\href{https://cumulocity.com/guides/reference/measurements/}{Cumulocity Measurements API}
}
\author{
Dmitriy Bolotov
}
