% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get the measurements for a device for a time period.}
\usage{
get_data(device_id, time_period)
}
\arguments{
\item{device_id}{The device id.}

\item{time_period}{The time period.}
}
\value{
R object with measurements.
}
\description{
Get the measurements for a device for a time period.
}
\details{
Get the measurements for a device for a time period.
}
\examples{
\dontrun{
get_data()
}
}
\references{
\href{https://cumulocity.com/guides/reference/events/}{Cumulocity Events API}
\href{https://cumulocity.com/guides/reference/measurements/}{Cumulocity Measurements API}
}
\author{
Dmitriy Bolotov
}
