% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_measurements.R
\name{get_measurements}
\alias{get_measurements}
\title{Get the measurements for a device for a time period.}
\usage{
get_measurements(device_id, date_from = NULL, date_to = NULL,
  page_size = 2000, abridged = TRUE, parse_datetime = TRUE)
}
\arguments{
\item{device_id}{The device id.}

\item{date_from}{The starting datetime.}

\item{date_to}{The ending datetime.}

\item{page_size}{The page size, set to 2000 (maximum) by default. Used
when at least one of the dates is NULL.}

\item{abridged}{If TRUE, exclude "self" and "source" fields from the result.}

\item{parse_datetime}{If TRUE, parse "time" field from char to POSIXlt.}
}
\value{
A \code{data.frame} with measurements.

Note that some columns in the returned data frame contain data frames themselves.
For example, \code{c8y_Mobile} is a \code{data.frame} of 11 variables.

If both \code{date_from} and \code{date_to} are present, \code{page_size}
is not used.

If \code{page_size} and both dates are NULL, the function will return up to 5 rows of data.
}
\description{
Get the measurements for a device for a time period.
}
\details{
Get the measurements for a device for a time period.

The parameter \code{page_size} is only used if \code{date_from}
or \code{date_to} are NULL, or both are NULL.
}
\examples{
\dontrun{
get_measurements(device_id)
}
}
\references{
\href{https://cumulocity.com/guides/reference/events/}{Cumulocity Events API}
\href{https://cumulocity.com/guides/reference/measurements/}{Cumulocity Measurements API}
}
\author{
Dmitriy Bolotov
}
