% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_devices.R
\name{list_devices}
\alias{list_devices}
\title{List available devices or group of devices for a tenant.}
\usage{
list_devices(page_size = NULL)
}
\arguments{
\item{page_size}{The page size.}
}
\value{
R object listing available devices.
}
\description{
List available devices or group of devices for a tenant.
}
\details{
List available devices or group of devices for a tenant.
}
\examples{
\dontrun{
list_devices()
}

}
\references{
\href{https://cumulocity.com/guides/reference/inventory/}{Cumulocity Inventory API}
}
\author{
Dmitriy Bolotov
}
