% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_devices.R
\name{list_devices}
\alias{list_devices}
\title{List available devices or group of devices for a tenant.}
\usage{
list_devices(param1)
}
\arguments{
\item{param1}{The first parameter.}
}
\value{
R object listing available devices.
}
\description{
List available devices or group of devices for a tenant.
}
\details{
List available devices or group of devices for a tenant.
}
\examples{
\dontrun{
print("example goes here")
}

}
\references{
\href{https://cumulocity.com/guides/reference/inventory/}{Cumulocity inventory API}
}
\author{
Dmitriy Bolotov
}
